Update README for htslib version
I get missing symbols with htslib 1.6 everything builds fine with the version of htslib that was contemporary at the last release of pbwt. I assume the incompatibility was introduced in a minor version tick. Its possible 1.4.X is compatible though. But 1.3.2 builds for me locally. A versioned submodule for the build might be a way forwards https://git-scm.com/docs/git-submodule